backend/app/requirements.txt
fastapi
uvicorn
qdrant-client
pydantic
requests
beautifulsoup4
tiktoken
openai
sqlite-vec
python-dotenv